+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED - 2025-March-20
 
+### Added
+
+- Support for setting ignores on XS-Check
+  - `XsManager.xs_check.ignores` is a `set` of warnings that will be ignored when processing
+  - `XsManager.xs_check.additional_args` is a `list` of arguments that will be appended to the XS-Check calls
+
 ### Fixed
 
 - Direct function calls in conditions and effects being added to the script provided to XS-Check